requirements.txt
pandas
numpy==1.19.3
pytesseract
requests
pillow
pandarallel
spacy
nltk
autocorrect
unidecode
scipy
sklearn
textblob
pyqt5-tools
matplotlib
